Use pthread_cond_broadcast() in queue.c
If there are multiple readers and multiple writers on the same queue, it's
possible for the pthread_cond_signal() that follows a write to be consumed by
another writer instead of one of the waiting readers.

This commit makes the write and read primitives in queue.c use
pthread_cond_broadcast() to avoid that issue.