Type-checking container_of macro
It is an error to invoke the container_of() macro with a
`ptr` parameter that is not of type pointer-to-`member`.
But the macro hides that error by ignoring the 'ptr' type.

Make the macro safer by requiring `ptr` to have the
correct type.

Also, rename it to er_container_of(), to avoid potential
name collisions with other container_of() macros.